# Profiling the self-hosted compiler