--- a/perfgraph.sh
+++ b/perfgraph.sh
@@ -7,9 +7,9 @@
 #   $ path/to/this/file ./program arg1 arg2
 
 dir=$(mktemp -d)
-trap 'rm -rvf "$dir"' EXIT
+trap 'rm -rf "$dir"' EXIT
 
 valgrind --tool=callgrind --callgrind-out-file=$dir/callgrind.out "$@"
 gprof2dot $dir/callgrind.out --format=callgrind --output=$dir/graph.gvz
-dot -Tpng $dir/graph.gvz -o $dir/image.png
-xdg-open $dir/image.png
+dot -Tpng $dir/graph.gvz -o $dir/perf.png
+mv -v $dir/perf.png perf.png