--- a/profiler
+++ b/profiler
@@ -1,11 +1,5 @@
 #!/bin/bash
 
-# profiling stuff
-#
-#   $ sudo apt install valgrind
-#   $ python3 -m pip install gprof2dot
-#   $ path/to/this/file ./program arg1 arg2
-
 dir=$(mktemp -d)
 trap 'rm -rf "$dir"' EXIT