--- a/self_hosted/c_output.oomph
+++ b/self_hosted/c_output.oomph
@@ -332,7 +332,6 @@ class FilePair(
                     "__Str___contains",
                     "__Str_center_pad",
                     "__Str_count",
-                    "__Str_ends_with",
                     "__Str_from_start_to_substring",
                     "__Str_get_utf8",
                     "__Str_left_pad",
@@ -344,7 +343,6 @@ class FilePair(
                     "__Str_right_pad",
                     "__Str_right_trim",
                     "__Str_split",
-                    "__Str_starts_with",
                     "__Str_trim",
                 ]:
                     # Class implemented in C, method in builtins.oomph