Benchmark: the `real` time it takes for the compiler to recompile itself
```
make oomph && time ./oomph --verbose -o oomph self_hosted/main.oomph
```

main branch: 1m0,171s
this branch: 0m23,079s
this branch and #142: 0m18,770s
pyoomph: 0m12,757s (so good lol)