@Akuli 2021-04-09
With this PR  and #142 (generating this takes a long time):

```
make oomph && time ./oomph --verbose -o /tmp/a self_hosted/main.oomph
```

![perf](https://user-images.githubusercontent.com/18505570/114169723-9b71a000-993a-11eb-9977-8c29a61c02dd.png)

Some observations:
- 45% emitting C code: could be improved by writing it to file instead of making huge string
- 31.99% emit_var: not as bad as it looks, includes emitting types which involves making lots of c code
- 12.12% `meth_List_LocalVariable_..._contains` + 11.07% `..._find_only`: having a mapping data structure would help (#198)
- 14.93% string splitting: I tried replacing `code.split("\n").first()` with `__slice_until_substring(code, "\n")` but didn't notice much difference
- 20.48% decref: the generated C code contains lots of unnecessary decrefs that the Oomph compiler could optimize away
- 6.55% memcpy: I don't understand how this gets called from things that should just create slices, such as `meth_Str_remove_prefix` and `oomph_get_first_char`