@Akuli 2021-04-09
running flamegraph

```
sudo apt install linux-perf
perf record -F 99 -a -g -- ./oomph --verbose -o /tmp/a self_hosted/main.oomph
git clone https://github.com/brendangregg/FlameGraph
cp perf.data FlameGraph/
cd FlameGraph
perf script | ./stackcollapse-perf.pl > out.perf-folded
./flamegraph.pl --width 10000 out.perf-folded > perf-kernel.svg
```

Can't upload svg to github :(